cdk_display_pipeline/requirements.txt
aws-cdk-lib>=2.140.0
constructs>=10.3.0
pytest>=8.0.0
